refactor: move `listable` into its own package - this package is then used by `list` and `changed` command, the CLI also has reference to its yarg CLI options
