--- packages/core/src/listable/index.ts
@@ -1,1 +1,0 @@
-export * from './listable-options';
