+++ packages/list/src/index.ts
@@ -1,3 +1,1 @@
 export * from './list-command';
-export * from './lib';
-export * from './types';
