--- packages/list/src/lib/index.ts
@@ -1,9 +1,0 @@
-import { listableOptions } from '@lerna-lite/core';
-
-import { listableFormat } from './listable-format';
-
-const listable = {
-  format: listableFormat,
-  options: listableOptions,
-};
-export { listable };
