+++ packages/listable/README.md
@@ -1,0 +1,16 @@
+[![License: MIT](https://img.shields.io/badge/License-MIT-yellow.svg)](https://opensource.org/licenses/MIT)
+[![npm](https://img.shields.io/npm/dy/@lerna-lite/listable?color=forest)](https://www.npmjs.com/package/@lerna-lite/listable)
+[![npm](https://img.shields.io/npm/v/@lerna-lite/listable.svg?logo=npm&logoColor=fff&label=npm)](https://www.npmjs.com/package/@lerna-lite/listable)
+
+## Lerna-Lite Listable utils used by `list` and `changed` commands
+
+#### @lerna-lite/listable
+
+Shared common code for the following optional commands:
+
+- [@lerna-lite/changed](https://github.com/ghiscoding/lerna-lite/tree/main/packages/changed)
+- [@lerna-lite/list](https://github.com/ghiscoding/lerna-lite/tree/main/packages/list)
+
+### Installation
+
+Follow the instruction provided in the main [README](https://github.com/ghiscoding/lerna-lite#installation) page.
