+++ packages/listable/src/index.ts
@@ -1,0 +1,8 @@
+import { listableFormat } from './lib/listable-format';
+import { listableOptions } from './lib/listable-options';
+
+const listable = {
+  format: listableFormat,
+  options: listableOptions,
+};
+export { listable };
